+++ CHANGELOG.md
@@ -8,6 +8,7 @@ dev
  ### Bug Fixes
 
  * ons-list-item: Fix expandable list item not working correctly for lists inside expandable content. ([#2485](https://github.com/OnsenUI/OnsenUI/issues/2485)).
+ * ons.platform.isIPhoneX: Support iPhone XS, XS Max, and XR. ([#2540](https://github.com/OnsenUI/OnsenUI/issues/2540)).
 
 2.10.4
 ---